+++ src/domain/locking/entities/activity-metadata.entity.ts
@@ -1,4 +1,11 @@
-import { ActivityMetadataSchema } from '@/domain/locking/entities/schemas/activity-metadata.schema';
+import { NumericStringSchema } from '@/validation/entities/schemas/numeric-string.schema';
 import { z } from 'zod';
 
 export type ActivityMetadata = z.infer<typeof ActivityMetadataSchema>;
+
+export const ActivityMetadataSchema = z.object({
+  campaignId: z.string(),
+  name: z.string(),
+  description: z.string(),
+  maxPoints: NumericStringSchema,
+});
