+++ src/domain/locking/entities/campaign.entity.ts
@@ -1,3 +1,4 @@
+import { ActivityMetadataSchema } from '@/domain/locking/entities/activity-metadata.entity';
 import { z } from 'zod';
 
 export type Campaign = z.infer<typeof CampaignSchema>;
@@ -9,5 +10,5 @@ export const CampaignSchema = z.object({
   periodStart: z.coerce.date(),
   periodEnd: z.coerce.date(),
   lastUpdated: z.coerce.date(),
-  // TODO: include 'activities' field once the structure is defined.
+  activities: z.array(ActivityMetadataSchema).nullish().default(null),
 });
