--- src/domain/locking/entities/schemas/activity-metadata.schema.ts
@@ -1,9 +1,0 @@
-import { NumericStringSchema } from '@/validation/entities/schemas/numeric-string.schema';
-import { z } from 'zod';
-
-export const ActivityMetadataSchema = z.object({
-  campaignId: z.string(),
-  name: z.string(),
-  description: z.string(),
-  maxPoints: NumericStringSchema,
-});
